refactor: 홈 데이터 변환과 타이머 렌더 범위 정리
